requirements.txt
torch
torchvision
flask
pillow
numpy
flask_cors